+++ cogs/review/review.py
@@ -66,6 +66,19 @@ class Review(rubbercog.Rubbercog):
         await utils.delete(ctx)
 
     @commands.check(acl.check)
+    @review.command(name="list", aliases=["available"])
+    async def review_list(self, ctx):
+        """Get list of reviewed subjects"""
+        subjects = set()
+        for r in repo_r.get_all_reviews():
+            subjects.add(r.subject)
+
+        if not len(subjects):
+            return await ctx.send(self.text.get("empty"))
+
+        await ctx.send(">>> " + ", ".join(f"`{s}`" for s in sorted(subjects)))
+
+    @commands.check(acl.check)
     @review.command(name="add", aliases=["update"])
     async def review_add(self, ctx, subject: str, mark: int, *, text: str):
         """Add a review
@@ -130,14 +143,14 @@ class Review(rubbercog.Rubbercog):
         return await ctx.send(self.text.get("removed"))
 
     @commands.check(acl.check)
-    @commands.group(name="sudo_subject")
-    async def sudo_subject(self, ctx):
+    @commands.group(name="subject")
+    async def subject(self, ctx):
         """Manage subjects"""
         await utils.send_help(ctx)
 
     @commands.check(acl.check)
-    @sudo_subject.command(name="add")
-    async def sudo_subject_add(self, ctx, subject: str, name: str, category: str):
+    @subject.command(name="add")
+    async def subject_add(self, ctx, subject: str, name: str, category: str):
         """Add subject
 
         subject: Subject code
@@ -153,8 +166,8 @@ class Review(rubbercog.Rubbercog):
         await ctx.send(self.text.get("subject_added"))
 
     @commands.check(acl.check)
-    @sudo_subject.command(name="update")
-    async def sudo_subject_update(self, ctx, subject: str, name: str, category: str):
+    @subject.command(name="update")
+    async def subject_update(self, ctx, subject: str, name: str, category: str):
         """Update subject
 
         subject: Subject code
@@ -170,8 +183,8 @@ class Review(rubbercog.Rubbercog):
         await ctx.send(self.text.get("subject_updated"))
 
     @commands.check(acl.check)
-    @sudo_subject.command(name="remove")
-    async def sudo_subject_remove(self, ctx, subject: str):
+    @subject.command(name="remove")
+    async def subject_remove(self, ctx, subject: str):
         """Remove subject
 
         subject: Subject code